fixes for condition_occurrence pk error

+++ src/main/resources/sqlserver/omop/OMOPCDM_sql_server_5.4_ddl.sql
@@ -396,7 +396,7 @@ CREATE TABLE CONDITION_ERA (
 --HINT DISTRIBUTE ON KEY (person_id)
 CREATE TABLE EPISODE (
 			episode_id bigint NOT NULL,
-			person_id bigint NOT NULL,
+			person_id integer NOT NULL,
 			episode_concept_id integer NOT NULL,
 			episode_start_date date NOT NULL,
 			episode_start_datetime datetime NULL,
